ai-service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
langchain==0.3.7
langchain-core==0.3.17
langchain-community==0.3.7
langchain-google-genai==2.0.3
langchain-text-splitters==0.3.2
chromadb==0.5.5
google-generativeai==0.8.3
pymysql==1.1.1
cryptography>=41.0.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
pydantic==2.9.2
langsmith==0.1.135
aiohttp==3.9.1
